+++ packages/web/src/core/dotlottie-player.js
@@ -875,18 +875,18 @@ var createDotLottiePlayerModule = (() => {
         C: function () {
           return 0;
         },
-        S: () => {},
-        P: () => {},
+        R: () => {},
+        O: () => {},
         T: function () {
           return 0;
         },
-        Q: () => {},
+        P: () => {},
         B: function () {},
-        R: () => {},
+        Q: () => {},
         U: () => {
           wa('');
         },
-        v: (a) => {
+        w: (a) => {
           var b = Ka[a];
           delete Ka[a];
           var c = b.Ua,
@@ -935,7 +935,7 @@ var createDotLottiePlayerModule = (() => {
             ];
           });
         },
-        J: () => {},
+        I: () => {},
         _: (a, b, c, d) => {
           b = P(b);
           N(a, {
@@ -1048,7 +1048,7 @@ var createDotLottiePlayerModule = (() => {
           });
         },
         Z: (a) => N(a, Cb),
-        x: (a, b, c, d) => {
+        y: (a, b, c, d) => {
           function e() {}
           b = P(b);
           e.values = {};
@@ -1087,7 +1087,7 @@ var createDotLottiePlayerModule = (() => {
             Ja: null,
           });
         },
-        F: (a, b, c, d, e, f) => {
+        v: (a, b, c, d, e, f) => {
           var h = ub(b, c);
           a = P(a);
           a = xb(a);
@@ -1239,7 +1239,7 @@ var createDotLottiePlayerModule = (() => {
             },
           });
         },
-        w: (a, b, c, d, e, f) => {
+        x: (a, b, c, d, e, f) => {
           Ka[a] = { name: P(b), Ua: T(c, d), Ka: T(e, f), Xa: [] };
         },
         j: (a, b, c, d, e, f, h, g, l, m) => {
@@ -1249,11 +1249,11 @@ var createDotLottiePlayerModule = (() => {
           b = P(b);
           N(a, { pb: !0, name: b, Ia: 0, fromWireType: () => {}, toWireType: () => {} });
         },
-        O: () => {
+        N: () => {
           Fa = !1;
           Vb = 0;
         },
-        K: () => {
+        J: () => {
           throw Infinity;
         },
         ca: (a, b, c, d) => {
@@ -1261,7 +1261,7 @@ var createDotLottiePlayerModule = (() => {
           b = Bb(b);
           return a(null, b, c, d);
         },
-        G: zb,
+        F: zb,
         ba: (a, b, c) => {
           var d = Rb(a, b),
             e = d.shift();
@@ -1292,7 +1292,7 @@ var createDotLottiePlayerModule = (() => {
           a = a.readValueFromPointer(b);
           return kb(a);
         },
-        L: (a, b) => {
+        K: (a, b) => {
           Tb[a] && (clearTimeout(Tb[a].id), delete Tb[a]);
           if (!b) return 0;
           var c = setTimeout(() => {
@@ -1302,7 +1302,7 @@ var createDotLottiePlayerModule = (() => {
           Tb[a] = { id: c, rb: b };
           return 0;
         },
-        M: (a, b, c, d) => {
+        L: (a, b, c, d) => {
           var e = new Date().getFullYear(),
             f = new Date(e, 0, 1).getTimezoneOffset();
           e = new Date(e, 6, 1).getTimezoneOffset();
@@ -1320,7 +1320,7 @@ var createDotLottiePlayerModule = (() => {
           e < f ? (I(a, c, 17), I(b, d, 17)) : (I(a, d, 17), I(b, c, 17));
         },
         fa: () => performance.now(),
-        N: (a) => {
+        M: (a) => {
           var b = x.length;
           a >>>= 0;
           if (2147483648 < a) return !1;
@@ -1364,10 +1364,10 @@ var createDotLottiePlayerModule = (() => {
         },
         s: () => 52,
         A: () => 52,
-        I: function () {
+        H: function () {
           return 70;
         },
-        z: (a, b, c, d) => {
+        S: (a, b, c, d) => {
           for (var e = 0, f = 0; f < c; f++) {
             var h = D[b >> 2],
               g = D[(b + 4) >> 2];
@@ -1391,12 +1391,12 @@ var createDotLottiePlayerModule = (() => {
         d: ic,
         e: jc,
         p: kc,
-        y: lc,
+        z: lc,
         b: mc,
         a: nc,
         g: oc,
         n: pc,
-        H: qc,
+        G: qc,
         V: Wb,
       },
       W = (function () {